+++ CHANGELOG.md
@@ -7,6 +7,23 @@
 - Cron job for CSV bulk trades
 - Get total invested value per stock
 
+## [0.0.6] - 2024-05-16
+### Added
+- Order services
+- Order API views
+- Order API permissions
+- Order API unit tests
+- Return test client `response` from `APITestCaseMixins`
+
+### Changed
+- Order manager queryset filter
+- Alter field `quantity` on `transaction` in `core`
+- Alter field `type` on `transaction` in `core`
+- Alter field `user` on `transaction` in `core`
+
+### Removed
+- Default `test.py` file
+
 ## [0.0.5] - 2024-05-15
 ### Added
 - Unit test for duplicated `code` field on `stock` in `core`